+++ Apps/CLI/Sources/PeekabooCLI/Commands/System/WindowCommand+Geometry.swift
@@ -82,7 +82,8 @@ extension WindowCommand {
                         isMainWindow: info.isMainWindow,
                         windowLevel: info.windowLevel,
                         alpha: info.alpha,
-                        index: info.index
+                        index: info.index,
+                        isOffScreen: info.isOffScreen
                     )
                 }